LayoutTests/inspector/unit-tests/async-test-suite-expected.txt

PASS: instantiating AsyncTestSuite requires name argument.
PASS: instantiating AsyncTestSuite requires string name argument.
PASS: instantiating AsyncTestSuite requires non-whitespace name argument.
PASS: instantiating AsyncTestSuite requires test harness argument.
PASS: should not be able to add empty test case.
PASS: should not be able to add non-object test case.
PASS: test case should require string name.
PASS: test case should require non-whitespace name.
PASS: test case should require test function.
PASS: should not be able to specify non-Function `setup` parameter.
PASS: should not be able to specify non-Function `setup` parameter.
PASS: should not be able to specify non-Function `setup` parameter.
PASS: should not be able to specify non-Function `teardown` parameter.
PASS: should not be able to specify non-Function `teardown` parameter.
PASS: should not be able to specify non-Function `teardown` parameter.
PASS: should not be able to run empty test suite.

== Running test suite: AsyncTestSuite.RunTwiceSuite
PASS: should not be able to run a test suite twice.
-- Running test case: DummyTest0

== Running test suite: AsyncTestSuite.SequentialExecution
PASS: AsyncTestSuite.RunTestCases() should return a Promise.
-- Running test case: DummyTest1
-- Running test case: DummyTest2
-- Running test case: DummyTest3
-- Running test case: FailingTest4
!! EXCEPTION: [object Object]
Stack Trace: (suppressed)
PASS: Promise from sequentialExecutionSuite.runTestCases() should reject when a test case fails.
PASS: Promise from sequentialExecutionSuite.runTestCases() should reject without altering its result value.
PASS: sequentialExecutionSuite should have executed four tests.
PASS: sequentialExecutionSuite should have passed three tests.
PASS: sequentialExecutionSuite should have failed 1 test.
PASS: sequentialExecutionSuite should have skipped zero tests.

== Running test suite: AsyncTestSuite.AbortOnFailure
-- Running test case: PassingTest5
-- Running test case: FailingTest6
!! EXCEPTION: {"token":666}
Stack Trace: (suppressed)
PASS: Promise from abortOnFailureSuite.runTestCases() should reject when a test case fails.
PASS: Promise from abortOnFailureSuite.runTestCases() should reject without altering its result value.
PASS: abortOnFailureSuite should have executed two tests.
PASS: abortOnFailureSuite should have passed one test.
PASS: abortOnFailureSuite should have failed one test.
PASS: abortOnFailureSuite should have skipped one test.

== Running test suite: AsyncTestSuite.SetupAndTeardown
-- Running test setup.
-- Running test case: TestWithSetupAndTeardown
PASS: Test should see side effects of running setup() action.
-- Running test teardown.
PASS: Teardown should see side effects of running setup() action.

-- Running test case: TestRunningAfterTeardown
PASS: Test should see side effects of previous test's teardown() action.
PASS: Promise from setupAndTeardownTestSuite.runTestCases() should resolve.

== Running test suite: AsyncTestSuite.SetupException
-- Running test setup.
!! EXCEPTION: 
Stack Trace: (suppressed)
PASS: Promise from setupExceptionTestSuite.runTestCases() should reject.

== Running test suite: AsyncTestSuite.SetupFailure
-- Running test setup.
!! EXCEPTION: undefined
Stack Trace: (suppressed)
PASS: Promise from setupFailureTestSuite.runTestCases() should reject.

== Running test suite: AsyncTestSuite.TeardownException
-- Running test case: TestWithExceptionDuringTeardown
-- Running test teardown.
!! EXCEPTION: 
Stack Trace: (suppressed)
PASS: Promise from teardownExceptionTestSuite.runTestCases() should reject.

== Running test suite: AsyncTestSuite.TeardownFailure
-- Running test case: TestWithExceptionDuringTeardown
-- Running test teardown.
!! EXCEPTION: undefined
Stack Trace: (suppressed)
PASS: Promise from teardownFailureTestSuite.runTestCases() should reject.

== Running test suite: AsyncTestSuite.AsyncFunctionSuccess
-- Running test case: AsyncFunctionSuccess
PASS: Promise from asyncFunctionSuccessTestSuite.runTestCases() should succeed.
PASS: Promise did evaluate the async test function.
PASS: Resolved value should be 42.

== Running test suite: AsyncTestSuite.AsyncFunctionExplicitFailure
-- Running test case: AsyncFunctionFailure
!! EXCEPTION: AsyncFunctionFailure Exception Message
Stack Trace: (suppressed)
PASS: Promise from asyncFunctionExplicitFailureTestSuite.runTestCases() should reject.
PASS: Promise did evaluate the async test function.
PASS: Rejected value should be thrown exception.

== Running test suite: AsyncTestSuite.AsyncFunctionRuntimeFailure
-- Running test case: AsyncFunctionFailure
!! EXCEPTION: undefined is not an object (evaluating '({}).x.x')
Stack Trace: (suppressed)
PASS: Promise from asyncFunctionRuntimeFailureTestSuite.runTestCases() should reject.
PASS: Promise did evaluate the async test function.
PASS: Rejected value should be a runtime exception.

== Running test suite: AsyncTestSuite.AsyncSetupAndAsyncTeardown
-- Running test setup.
-- Running test case: TestWithSetupAndTeardown
PASS: Test should see side effects of running setup() action.
-- Running test teardown.
PASS: Teardown should see side effects of running setup() action.

-- Running test case: TestRunningAfterTeardown
PASS: Test should see side effects of previous test's teardown() action.
PASS: Promise from asyncSetupAndAsyncTeardownTestSuite.runTestCases() should resolve.

== Running test suite: AsyncTestSuite.AsyncSetupExplicitFailure
-- Running test case: AsyncFunctionFailure
!! EXCEPTION: AsyncFunctionFailure Exception Message
Stack Trace: (suppressed)
PASS: Promise from asyncSetupExplicitFailureTestSuite.runTestCases() should reject.
PASS: Promise did evaluate the async setup function.
PASS: Rejected value should be thrown exception.

== Running test suite: AsyncTestSuite.AsyncSetupRuntimeFailure
-- Running test setup.
!! EXCEPTION: undefined is not an object (evaluating '({}).x.x')
Stack Trace: (suppressed)
PASS: Promise from asyncSetupRuntimeFailureTestSuite.runTestCases() should reject.
PASS: Promise did evaluate the async setup function.
PASS: Rejected value should be a runtime exception.

== Running test suite: AsyncTestSuite.AsyncTeardownExplicitFailure
-- Running test case: AsyncFunctionFailure
-- Running test teardown.
!! EXCEPTION: AsyncFunctionFailure Exception Message
Stack Trace: (suppressed)
PASS: Promise from asyncTeardownExplicitFailureTestSuite.runTestCases() should reject.
PASS: Promise did evaluate the async teardown function.
PASS: Rejected value should be thrown exception.

== Running test suite: AsyncTestSuite.AsyncTeardownRuntimeFailure
-- Running test case: AsyncFunctionFailure
-- Running test teardown.
!! EXCEPTION: undefined is not an object (evaluating '({}).x.x')
Stack Trace: (suppressed)
PASS: Promise from asyncTeardownRuntimeFailureTestSuite.runTestCases() should reject.
PASS: Promise did evaluate the async teardown function.
PASS: Rejected value should be a runtime exception.
